requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
faiss-cpu==1.11.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
langsmith==0.3.43
numpy==2.2.6
openai==1.82.1
orjson==3.10.18
packaging==24.2
pinecone==7.0.2
pinecone-plugin-assistant==1.6.1
python-dotenv==1.0.0
pinecone-plugin-interface==0.0.7
pip==24.0
pydantic==2.11.5
pydantic-core==2.33.2
python-dateutil==2.9.0.post0
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.1
urllib3==2.4.0
zstandard==0.23.0
